Factor out JavaScript template from Python script
The template is over 350 lines of JavaScript, which is hard to work with
embedded in a multiline Python string. Placing it in a separate file
will make future changes _much_ easier to implement and review.
